+++ admin/class-admin-init.php
@@ -83,7 +83,8 @@ class WPSEO_Admin_Init {
 		$current_url   = ( is_ssl() ? 'https://' : 'http://' );
 		$current_url  .= sanitize_text_field( $_SERVER['SERVER_NAME'] ) . sanitize_text_field( $_SERVER['REQUEST_URI'] );
 		$customize_url = add_query_arg( array(
-			'url' => urlencode( $current_url ),
+			'autofocus[control]' => 'blogdescription',
+			'url'                => urlencode( $current_url ),
 		), wp_customize_url() );
 
 		$info_message = sprintf(