+++ admin/views/tabs/metas/taxonomies.php
@@ -44,8 +44,13 @@ if ( is_array( $taxonomies ) && $taxonomies !== array() ) {
 			$taxonomies_help->get_button_html() . $taxonomies_help->get_panel_html()
 		);
 
-		$yform->textinput( 'title-tax-' . $tax->name, __( 'Title template', 'wordpress-seo' ), 'template taxonomy-template' );
-		$yform->textarea( 'metadesc-tax-' . $tax->name, __( 'Meta description template', 'wordpress-seo' ), array( 'class' => 'template taxonomy-template' ) );
+		// Determine the page type for the term, this is needed for the recommended replacement variables.
+		$recommended_replace_vars = new WPSEO_Admin_Recommended_Replace_Vars();
+		$page_type                = $recommended_replace_vars->determine_for_term( $tax->name );
+
+		$editor = new WPSEO_Replacevar_Editor( $yform, 'title-tax-' . $tax->name, 'metadesc-tax-' . $tax->name, $page_type );
+		$editor->render();
+
 		if ( $tax->name !== 'post_format' ) {
 			/* translators: %1$s expands to Yoast SEO */
 			$yform->show_hide_switch( 'display-metabox-tax-' . $tax->name, sprintf( __( '%1$s Meta Box', 'wordpress-seo' ), 'Yoast SEO' ) );