+++ js/tests/components/SnippetPreviewSection.test.js
@@ -1,6 +1,7 @@
-import { mount } from "enzyme";
-import SnippetPreviewSection from "../../src/components/SnippetPreviewSection";
 import React from "react";
+import { mount } from "enzyme";
+
+import SnippetPreviewSection from "../../src/components/SnippetPreviewSection";
 
 jest.mock( "../../src/containers/SnippetEditor", () => {
 	return () => {
@@ -14,7 +15,11 @@ jest.mock( "yoast-components", () => {
 
 describe( "SnippetPreviewSection", () => {
 	it( "renders the snippet editor inside of it", () => {
-		const tree = mount( <SnippetPreviewSection baseUrl="http://example.org" /> );
+		const tree = mount(
+			<SnippetPreviewSection
+				title="Snippet editor"
+				icon="eye" />
+		);
 
 		expect( tree ).toMatchSnapshot();
 	} );